fetch weather data from maps to result activity using intent

+++ app/src/main/java/com/capstone/sopanfinder/api/ApiService.kt
@@ -37,12 +37,12 @@ interface ApiService {
     @POST("predict_text")
     fun fetchResult(
         @Field("date") date: List<String>,
-        @Field("pv_demand") pv_demand: List<Int>,
-        @Field("temp") temp: List<Double>,
-        @Field("wind_speed") wind_speed: List<Double>,
-        @Field("rain_1h") rain_1h: List<Double>,
-        @Field("snow_1h") snow_1h: List<Double>,
-        @Field("clouds_all") clouds_all: List<Int>,
-        @Field("percip_1h") percip_1h: List<Double>
+        @Field("pv_demand") pv_demand: List<Float>,
+        @Field("temp") temp: List<Float>,
+        @Field("wind_speed") wind_speed: List<Float>,
+        @Field("rain_1h") rain_1h: List<Float>,
+        @Field("snow_1h") snow_1h: List<Float>,
+        @Field("clouds_all") clouds_all: List<Float>,
+        @Field("percip_1h") percip_1h: List<Float>
     ): Call<SopanResponse>
 }
